Bring PathInternal methods into SqlFileStream, remove the PathInternal file.
